extract ResourcesModule
