+++ integration/src/test/scala/org/knora/webapi/core/LayersTestLive.scala
@@ -45,6 +45,7 @@ import org.knora.webapi.slice.ontology.domain.service.CardinalityService
 import org.knora.webapi.slice.ontology.repo.service.OntologyCache
 import org.knora.webapi.slice.resourceinfo.ResourceInfoLayers
 import org.knora.webapi.slice.resourceinfo.api.ResourceInfoEndpoints
+import org.knora.webapi.slice.resources.ResourcesModule
 import org.knora.webapi.slice.resources.api.ResourcesApiModule
 import org.knora.webapi.slice.resources.repo.service.ResourcesRepo
 import org.knora.webapi.slice.resources.repo.service.ResourcesRepoLive
@@ -124,6 +125,7 @@ object LayersTestLive { self =>
       ResourceInfoLayers.live,
       ResourceUtilV2Live.layer,
       ResourcesApiModule.layer,
+      ResourcesModule.layer,
       ResourcesRepoLive.layer,
       ResourcesResponderV2.layer,
       SearchApiRoutes.layer,
