+++ integration/src/test/scala/org/knora/webapi/core/LayersTestMock.scala
@@ -45,6 +45,7 @@ import org.knora.webapi.slice.ontology.domain.service.OntologyTriplestoreHelpers
 import org.knora.webapi.slice.ontology.repo.service.OntologyCache
 import org.knora.webapi.slice.resourceinfo.ResourceInfoLayers
 import org.knora.webapi.slice.resourceinfo.api.ResourceInfoEndpoints
+import org.knora.webapi.slice.resources.ResourcesModule
 import org.knora.webapi.slice.resources.api.ResourcesApiModule
 import org.knora.webapi.slice.resources.repo.service.ResourcesRepo
 import org.knora.webapi.slice.resources.repo.service.ResourcesRepoLive
@@ -132,6 +133,7 @@ object LayersTestMock { self =>
       ResourceInfoLayers.live,
       ResourceUtilV2Live.layer,
       ResourcesApiModule.layer,
+      ResourcesModule.layer,
       ResourcesRepoLive.layer,
       ResourcesResponderV2.layer,
       SearchApiRoutes.layer,
