+++ webapi/src/main/scala/org/knora/webapi/responders/v2/ResourcesResponderV2.scala
@@ -63,7 +63,6 @@ import org.knora.webapi.slice.common.repo.rdf.Vocabulary
 import org.knora.webapi.slice.common.repo.rdf.Vocabulary.KnoraBase as KB
 import org.knora.webapi.slice.ontology.domain.service.IriConverter
 import org.knora.webapi.slice.ontology.domain.service.OntologyRepo
-import org.knora.webapi.slice.resources.api.ResourceMetadataDto
 import org.knora.webapi.slice.resources.api.model.GraphDirection
 import org.knora.webapi.slice.resources.api.model.VersionDate
 import org.knora.webapi.slice.resources.repo.service.ResourcesRepo
@@ -174,77 +173,6 @@ final case class ResourcesResponderV2(
 
   def createResource(createResource: CreateResourceRequestV2): Task[ReadResourcesSequenceV2] =
     createHandler(createResource)
-
-  def getResourcesMetadata(project: KnoraProject): Task[List[ResourceMetadataDto]] = {
-    val (
-      classIriVar,
-      creationDateVar,
-      creatorIriVar,
-      deleteDateVar,
-      labelVar,
-      lastModificationDateVar,
-      resourceIriVar,
-    ) = ("classIri", "createdAt", "creator", "deletedAt", "label", "modifiedAt", "resourceIri")
-
-    val selectPattern = SparqlBuilder
-      .select(
-        variable(classIriVar),
-        variable(creationDateVar),
-        variable(creatorIriVar),
-        variable(deleteDateVar),
-        variable(labelVar),
-        variable(lastModificationDateVar),
-        variable(resourceIriVar),
-      )
-      .distinct()
-
-    val projectGraph = projectService.getDataGraphForProject(project)
-    val wherePattern =
-      variable(resourceIriVar)
-        .isA(variable(classIriVar))
-        .andHas(KB.creationDate, variable(creationDateVar))
-        .andHas(KB.attachedToUser, variable(creatorIriVar))
-        .andHas(RDFS.LABEL, variable(labelVar))
-        .and(variable(resourceIriVar).has(KB.lastModificationDate, variable(lastModificationDateVar)).optional())
-        .and(variable(resourceIriVar).has(KB.deleteDate, variable(deleteDateVar)).optional())
-        .from(Rdf.iri(projectGraph.value))
-
-    val classConstraintPattern =
-      variable(classIriVar).has(PropertyPathBuilder.of(RDFS.SUBCLASSOF).zeroOrMore().build(), KB.Resource)
-
-    val query = Queries
-      .SELECT(selectPattern)
-      .where(wherePattern, classConstraintPattern)
-      .prefix(prefix(KB.NS), prefix(RDFS.NS))
-
-    def throwEx(field: String): Nothing = throw new InconsistentRepositoryDataException(
-      s"Resource metadata query for project ${project.shortcode} returned inconsistent data for $field",
-    )
-    for {
-      _ <- ZIO.logInfo(s"QUERY: \n ${query.getQueryString}")
-      rows <- triplestore
-                .select(query)
-                .map(_.results.bindings)
-                .timed
-                .flatMap((d, s) => ZIO.logInfo(s"Query took ${d.toMillis} ms and returned ${s.size} rows").ignore.as(s))
-      meta <- ZIO
-                .attempt(rows.map { row =>
-                  val resourceIri = row.rowMap.getOrElse(resourceIriVar, throwEx(resourceIriVar))
-                  val classIri    = row.rowMap.getOrElse(classIriVar, throwEx(classIriVar))
-                  val creatorIri  = row.rowMap.getOrElse(creatorIriVar, throwEx(creatorIriVar))
-                  val label       = row.rowMap.getOrElse(labelVar, throwEx(labelVar))
-                  val createdAt   = row.rowMap.get(creationDateVar).map(Instant.parse).getOrElse(throwEx(creationDateVar))
-                  val deletedAt   = row.rowMap.get(deleteDateVar).map(Instant.parse)
-                  val lastModAt   = row.rowMap.get(lastModificationDateVar).map(Instant.parse)
-                  val arkUrl      = resourceIri.toSmartIri.fromResourceIriToArkUrl()
-                  ResourceMetadataDto(classIri, resourceIri, arkUrl, label, creatorIri, createdAt, lastModAt, deletedAt)
-                })
-                .timed
-                .flatMap((d, m) =>
-                  ZIO.logInfo(s"Mapping took ${d.toMillis} ms and returned ${m.size} rows").ignore.as(m.toList),
-                )
-    } yield meta
-  }
 
   /**
    * If resource has already been modified, make sure that its lastModificationDate is given in the request body.
